# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sustainable AI Lifecycle Auditor API",
    "description": "Estimates lifecycle energy, carbon, and hardware footprint of AI systems.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ledger": {
      "post": {
        "summary": "Calculate Ledger",
        "description": "Compute dual lifecycle ledgers (operational and full_lifecycle)\nfor a given AI system inventory.",
        "operationId": "calculate_ledger_api_v1_ledger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/check": {
      "post": {
        "summary": "Check Efficiency Claim",
        "description": "Audit an efficiency claim against baseline and subject ledgers\nusing the deterministic multi-tier verification cascade.",
        "operationId": "check_efficiency_claim_api_v1_claims_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimToLedgerMapping"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verdict"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/extract": {
      "post": {
        "summary": "Extract Claim",
        "description": "Extract a structured Claim object from free-text marketing copy,\nresearch papers, or model cards.",
        "operationId": "extract_claim_api_v1_claims_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommend": {
      "post": {
        "summary": "Recommend Architectures",
        "description": "Generate alternative architectures, evaluate full lifecycle ledgers,\nfilter constraints, and compute Pareto frontier.",
        "operationId": "recommend_architectures_api_v1_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParetoResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountingMethod": {
        "type": "string",
        "enum": [
          "location",
          "market"
        ],
        "title": "AccountingMethod"
      },
...
      "Boundary": {
        "type": "string",
        "enum": [
          "operational",
          "full_lifecycle"
        ],
        "title": "Boundary"
      },
...
      "Claim": {
        "properties": {
          "claim_text": {
            "type": "string",
            "title": "Claim Text"
          },
          "metric": {
            "type": "string",
            "enum": [
              "energy",
              "carbon",
              "latency",
              "cost"
            ],
            "title": "Metric"
          },
          "claimed_change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed Change Pct"
          },
          "baseline_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Ref"
          },
          "subject_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Ref"
          },
          "boundary_stated": {
            "$ref": "#/components/schemas/BoundaryStated"
          },
          "hardware": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "accuracy_change_reported": {
            "type": "boolean",
            "title": "Accuracy Change Reported",
            "default": false
          },
          "accuracy_change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Change Pct"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "vendor",
              "paper",
              "model_card",
              "blog",
              "unknown"
            ],
            "title": "Source Type",
            "default": "unknown"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "extraction_source": {
            "type": "string",
            "enum": [
              "gemini",
              "rules_fallback",
              "user_specified"
            ],
            "title": "Extraction Source",
            "default": "user_specified"
          }
        },
        "type": "object",
        "required": [
          "claim_text",
          "metric",
          "boundary_stated"
        ],
        "title": "Claim",
        "description": "A structured efficiency claim extracted from free text or a model card.\n\nFields left None mean the information was absent or ambiguous in the source.\nNever guess missing boundary fields: leave them unspecified / None."
      },
      "ClaimToLedgerMapping": {
        "properties": {
          "claim": {
            "$ref": "#/components/schemas/Claim"
          },
          "baseline_ledger": {
            "$ref": "#/components/schemas/LifecycleLedger-Input"
          },
          "subject_ledger": {
            "$ref": "#/components/schemas/LifecycleLedger-Input"
          },
          "baseline_ledger_full": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LifecycleLedger-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "subject_ledger_full": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LifecycleLedger-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "claim",
          "baseline_ledger",
          "subject_ledger"
        ],
        "title": "ClaimToLedgerMapping",
        "description": "User-supplied mapping that links a Claim to computed LifecycleLedgers.\n\nCan supply both operational ledgers (baseline_ledger, subject_ledger)\nand full-lifecycle ledgers (baseline_ledger_full, subject_ledger_full).\nIf full-lifecycle ledgers are omitted, the checker uses baseline_ledger\nand subject_ledger under their native boundary."
      },
      "ExtractClaimRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ExtractClaimRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LedgerEngineConfig": {
        "properties": {
          "pue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pue"
          },
          "hardware_lifetime_years": {
            "type": "number",
            "title": "Hardware Lifetime Years",
            "default": 4.0
          },
          "utilization": {
            "type": "number",
            "title": "Utilization",
            "default": 1.0
          },
          "accounting": {
            "$ref": "#/components/schemas/AccountingMethod",
            "default": "location"
          },
          "n_mc_samples": {
            "type": "integer",
            "title": "N Mc Samples",
            "default": 1000
          },
          "mc_seed": {
            "type": "integer",
            "title": "Mc Seed",
            "default": 42
          },
          "requests_per_day": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requests Per Day"
          }
        },
        "type": "object",
        "title": "LedgerEngineConfig",
        "description": "Overridable assumptions for the ledger engine.\nDefaults match the recommended values from their respective sources."
      },
      "LedgerRequest": {
        "properties": {
          "inventory": {
            "$ref": "#/components/schemas/SystemInventory"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LedgerEngineConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "inventory"
        ],
        "title": "LedgerRequest"
      },
      "LedgerResponse": {
        "properties": {
          "operational": {
            "$ref": "#/components/schemas/LifecycleLedger-Output"
          },
          "full_lifecycle": {
            "$ref": "#/components/schemas/LifecycleLedger-Output"
          }
        },
        "type": "object",
        "required": [
          "operational",
          "full_lifecycle"
        ],
        "title": "LedgerResponse"
      },
...
      "ParetoResult": {
        "properties": {
          "boundary": {
            "$ref": "#/components/schemas/Boundary"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RecommenderCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "pareto_front_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pareto Front Labels"
          },
          "break_even_traffic_requests": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Break Even Traffic Requests"
          },
          "winner_operational": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Operational"
          },
          "winner_full_lifecycle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Full Lifecycle"
          },
          "winner_changed": {
            "type": "boolean",
            "title": "Winner Changed",
            "default": false
          },
          "qualifying_candidates_count": {
            "type": "integer",
            "title": "Qualifying Candidates Count",
            "default": 0
          },
          "binding_constraint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binding Constraint"
          },
          "rankings_operational": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rankings Operational",
            "description": "Ranked candidate labels sorted by operational carbon (lowest to highest)."
          },
          "rankings_full_lifecycle": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rankings Full Lifecycle",
            "description": "Ranked candidate labels sorted by full lifecycle carbon (lowest to highest)."
          },
          "pareto_points_operational": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pareto Points Operational",
            "description": "Non-dominated Pareto frontier labels under operational boundary."
          },
          "pareto_points_full_lifecycle": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pareto Points Full Lifecycle",
            "description": "Non-dominated Pareto frontier labels under full lifecycle boundary."
          },
          "break_even_operational_requests": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Break Even Operational Requests",
            "description": "Break-even request volume under operational boundary."
          },
          "break_even_full_lifecycle_requests": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Break Even Full Lifecycle Requests",
            "description": "Break-even request volume under full lifecycle boundary."
          }
        },
        "type": "object",
        "required": [
          "boundary",
          "candidates",
          "pareto_front_labels"
        ],
        "title": "ParetoResult",
        "description": "Pareto analysis output for the recommender screen across both boundaries."
      },
...
      "RecommendRequest": {
        "properties": {
          "inventory": {
            "$ref": "#/components/schemas/SystemInventory"
          },
          "boundary": {
            "$ref": "#/components/schemas/Boundary",
            "default": "operational"
          },
          "accuracy_floor": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Floor"
          }
        },
        "type": "object",
        "required": [
          "inventory"
        ],
        "title": "RecommendRequest"
      },
...
      "SystemInventory": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "training": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainingConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "inference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "region": {
            "type": "string",
            "title": "Region",
            "description": "Primary deployment region. Used for grid intensity lookup.",
            "default": "global"
          },
          "retraining_frequency_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retraining Frequency Days",
            "description": "How often the model is retrained. None means no retraining."
          },
          "storage_tb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Tb",
            "description": "Storage footprint in TB (model weights, data, checkpoints)."
          },
          "network_requests_per_day": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network Requests Per Day",
            "description": "Daily request count for network transfer estimation."
          },
          "accuracy_floor": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Floor",
            "description": "Minimum acceptable accuracy (0-1). Recommender excludes configs below this."
          },
          "distillation_training_hours": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distillation Training Hours",
            "description": "Customizable distillation training run hours (default 50.0h)."
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "Free-text notes from config parser.",
            "default": ""
          }
        },
        "type": "object",
        "title": "SystemInventory",
        "description": "Structured description of an AI system to audit.\n\nFields not present in a parsed config file stay None and are shown\nto the user as 'please confirm' in the UI."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "properties": {
          "label": {
            "$ref": "#/components/schemas/VerdictLabel"
          },
          "resolved_by": {
            "type": "string",
            "enum": [
              "rules",
              "classifier",
              "llm"
            ],
            "title": "Resolved By"
          },
          "reason_chain": {
            "items": {
              "$ref": "#/components/schemas/ReasonStep"
            },
            "type": "array",
            "title": "Reason Chain"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Confidence [0,1]. None means the check is deterministic."
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "label",
          "resolved_by",
          "reason_chain"
        ],
        "title": "Verdict"
      },
      "VerdictLabel": {
        "type": "string",
        "enum": [
          "supported",
          "contradicted",
          "suspicious_boundary_shift",
          "insufficient_evidence"
        ],
        "title": "VerdictLabel"
      }
    }
  }
}